+++ apps/mobile/src/app/(home,search,profile,notifications,settings)/_layout.tsx
@@ -5,6 +5,7 @@ import { useTranslations } from 'use-intl'
 
 import { Icon } from '~/components/common/icon'
 import { IconButton } from '~/components/common/icon-button'
+import { Pressable } from '~/components/common/pressable'
 import { Text } from '~/components/common/text'
 import { drawer, HomeDrawer } from '~/components/home/drawer'
 import { StackHeader } from '~/components/navigation/stack-header'
@@ -18,6 +19,7 @@ import { type Undefined } from '~/types'
 
 import { type HomeParams } from '.'
 import { type CommunityParams } from './communities/[name]'
+import { type MessageParams } from './messages/[id]'
 import { type PostParams } from './posts/[id]'
 import { type SignInParams } from './sign-in'
 import { type UserParams } from './users/[name]'
@@ -296,6 +298,35 @@ function StackLayout({ children }: PropsWithChildren) {
           presentation: 'modal',
           title: t('posts.reply.title'),
         }}
+      />
+
+      <Stack.Screen
+        name="messages/[id]"
+        options={({ route }) => ({
+          headerTitle: () => {
+            const { user } = route.params as MessageParams
+
+            if (user) {
+              return (
+                <Pressable
+                  height="8"
+                  justify="center"
+                  onPress={() => {
+                    router.push({
+                      params: {
+                        name: user,
+                      },
+                      pathname: '/users/[name]',
+                    })
+                  }}
+                  px="3"
+                >
+                  <Text weight="bold">{user}</Text>
+                </Pressable>
+              )
+            }
+          },
+        })}
       />
 
       <Stack.Screen