+++ apps/mobile/src/components/messages/card.tsx
@@ -1,0 +1,48 @@
+import { isToday } from 'date-fns'
+import { createStyleSheet, useStyles } from 'react-native-unistyles'
+import { useFormatter } from 'use-intl'
+
+import { type Message } from '~/types/message'
+
+import { Markdown } from '../common/markdown'
+import { Text } from '../common/text'
+import { View } from '../common/view'
+
+type Props = {
+  message: Message
+  userId?: string
+}
+
+export function MessageCard({ message, userId }: Props) {
+  const f = useFormatter()
+
+  const { styles } = useStyles(stylesheet)
+
+  const self = message.author === userId
+
+  return (
+    <View align={self ? 'end' : 'start'} gap="1" self={self ? 'end' : 'start'}>
+      <View px="2" py="1" style={styles.content(self)}>
+        <Markdown recyclingKey={message.id} variant="comment">
+          {message.body}
+        </Markdown>
+      </View>
+
+      <Text highContrast={false} size="1" tabular>
+        {f.dateTime(message.createdAt, {
+          dateStyle: !isToday(message.createdAt) ? 'short' : undefined,
+          timeStyle: 'short',
+        })}
+      </Text>
+    </View>
+  )
+}
+
+const stylesheet = createStyleSheet((theme) => ({
+  content: (self: boolean) => ({
+    backgroundColor: self ? theme.colors.accent.ui : theme.colors.gray.ui,
+    borderCurve: 'continuous',
+    borderRadius: theme.space[2],
+    maxWidth: '90%',
+  }),
+}))