+++ apps/mobile/src/types/message.ts
@@ -1,0 +1,9 @@
+export type Message = {
+  author: string
+  body: string
+  createdAt: Date
+  id: string
+  new: boolean
+  replies?: Array<Message>
+  subject: string
+}